Documented every scene controller class and the entry point. Also moved the controllers one folder up.
