+++ src/main/java/me/pr0crustes/Starter.java
@@ -4,31 +4,49 @@ import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
-import me.pr0crustes.frontend.gui.scene.menu.MenuController;
+import me.pr0crustes.frontend.gui.scene.MenuController;
 
+/**
+ * Class that is the entry point of the program.
+ * Since it is a JavaFX app, extends Application.
+ * @see Application
+ */
 public class Starter extends Application {
 
-    public static Stage mainStage; // Singleton
+    /**
+     * A Stage singleton. This will come in hand when we need to show file select / file save dialogs.
+     */
+    public static Stage mainStage;
 
-    private static final String programTitle = "PDF-Toolkit";
-
+    /**
+     * JavaFX Application start method.
+     * Launchs the main window.
+     * Creates a MenuController.
+     * @param primaryStage the JavaFX primaryStage.
+     * @see Application
+     * @see MenuController
+     */
     @Override
     public void start(Stage primaryStage) {
         Starter.mainStage = primaryStage;
 
         AnchorPane root = new AnchorPane();
-
         Scene scene = new Scene(root, 600, 400);
 
         primaryStage.setScene(scene);
-        primaryStage.setTitle(Starter.programTitle);
+        primaryStage.setTitle("PDF-Toolkit");
         primaryStage.setResizable(false);
         primaryStage.show();
 
         new MenuController(root);
     }
 
+    /**
+     * Main, entry point!
+     * @param args something...
+     */
     public static void main(String[] args) {
+        // Keep this line. Works better with it. see https://pdfbox.apache.org/2.0/getting-started.html
         System.setProperty("sun.java2d.cmm", "sun.java2d.cmm.kcms.KcmsServiceProvider");
         launch(args);
     }
