+++ src/main/java/me/pr0crustes/frontend/gui/scene/MenuController.java
@@ -1,4 +1,4 @@
-package me.pr0crustes.frontend.gui.scene.menu;
+package me.pr0crustes.frontend.gui.scene;
 
 import javafx.geometry.Pos;
 import javafx.scene.control.Button;
@@ -8,29 +8,40 @@ import javafx.scene.layout.Pane;
 import javafx.scene.layout.StackPane;
 import me.pr0crustes.frontend.gui.classes.PassiveController;
 import me.pr0crustes.frontend.gui.classes.layout.NodeFactory;
-import me.pr0crustes.frontend.gui.scene.convert.ConvertController;
-import me.pr0crustes.frontend.gui.scene.crop.CropController;
-import me.pr0crustes.frontend.gui.scene.insert.InsertController;
-import me.pr0crustes.frontend.gui.scene.merge.MergeController;
-import me.pr0crustes.frontend.gui.scene.quality.QualityController;
 
+/**
+ * MenuController is the root controller.
+ * Creates all the buttons for others tabs.
+ * Extends PassiveController.
+ * @see PassiveController
+ */
 public class MenuController extends PassiveController {
 
+    /**
+     * Just a constructor that calls super.
+     * @param pane the Pane the GUI should be drawn on.
+     */
     public MenuController(Pane pane) {
         super(pane);
     }
 
+    /**
+     * Implementation of setupGUI.
+     * @param pane the pane the GUI should be made on.
+     * @see me.pr0crustes.frontend.gui.classes.Setup
+     */
     @Override
     public void setupGUI(Pane pane) {
         SplitPane splitPane = this.createSplitPane();
-
         NodeFactory.bindToParent(splitPane, pane);
-
         pane.getChildren().add(splitPane);
     }
 
+    /**
+     * Method that creates the GUI splitPane and returns.
+     * @return a SplitPane with everything on it.
+     */
     private SplitPane createSplitPane() {
-
         SplitPane splitPane = new SplitPane();
 
         StackPane stackPaneRight = new StackPane();
@@ -42,6 +53,11 @@ public class MenuController extends PassiveController {
         return splitPane;
     }
 
+    /**
+     * Method that creates a StackPane with all the buttons.
+     * @param stackPaneRight the right stackPane. This is needed because it will be passed to all the others controllers.
+     * @return the stackPane that will be on the left.
+     */
     private StackPane createLeftStackPane(StackPane stackPaneRight) {
 
         Button[] buttons = {
